fix: cast _SC_NPROCESSORS_ONLN to Int32 for Linux compatibility

+++ Sources/IO Primitives/IO.Platform.swift
@@ -30,7 +30,7 @@ extension IO.Platform {
             GetSystemInfo(&info)
             return Int(info.dwNumberOfProcessors)
         #else
-            let count = sysconf(_SC_NPROCESSORS_ONLN)
+            let count = sysconf(Int32(_SC_NPROCESSORS_ONLN))
             return count > 0 ? Int(count) : 1
         #endif
     }
